+++ thoth-client/src/queries.rs
@@ -13,7 +13,7 @@ use uuid::Uuid;
 
 #[derive(GraphQLQuery)]
 #[graphql(
-    schema_path = "assets/schema.json",
+    schema_path = "assets/schema.graphql",
     query_path = "assets/queries.graphql",
     response_derives = "Debug,Clone,Deserialize,Serialize,PartialEq",
     variables_derives = "Debug,PartialEq"
@@ -40,7 +40,7 @@ impl fmt::Display for work_query::PublicationType {
 
 #[derive(GraphQLQuery)]
 #[graphql(
-    schema_path = "assets/schema.json",
+    schema_path = "assets/schema.graphql",
     query_path = "assets/queries.graphql",
     response_derives = "Debug,Clone,Deserialize,Serialize,PartialEq",
     variables_derives = "Debug,PartialEq"
@@ -49,7 +49,7 @@ pub struct WorksQuery;
 
 #[derive(GraphQLQuery)]
 #[graphql(
-    schema_path = "assets/schema.json",
+    schema_path = "assets/schema.graphql",
     query_path = "assets/queries.graphql",
     response_derives = "Debug,Clone,Deserialize,Serialize,PartialEq",
     variables_derives = "Debug,PartialEq"